requirements.txt
tensorflow==2.10.1
medmnist==2.2.2
numpy==1.23.5
scikit-learn==1.3.2
matplotlib==3.7.3
pyyaml==6.0.2
pytest==7.4.4